model/requirements.txt
transformers
torch
pandas
pyyaml
cassandra-driver